Moved transform.hpp into components as it is now a component
Fixed path separator for Linux platforms.
