+++ include/components/transforms.hpp
@@ -1,0 +1,31 @@
+#pragma once
+
+#include <glm/glm.hpp>
+#include <glm/gtx/quaternion.hpp>
+#include <glm/gtc/quaternion.hpp>
+
+namespace tec {
+	static glm::vec3 FORWARD_VECTOR(0.0f, 0.0f, -1.0f);
+	static glm::vec3 UP_VECTOR(0.0f, 1.0f, 0.0f);
+	static glm::vec3 RIGHT_VECTOR(1.0f, 0.0f, 0.0f);
+
+	struct Position {
+		// Translates by amount.
+		void Translate(const glm::vec3 amount);
+
+		// Translates by amount in direction orientation.
+		void Translate(const glm::vec3 amount, const glm::quat orientation);
+
+		glm::vec3 value;
+	};
+
+	struct Orientation {
+		void Rotate(const glm::vec3 amount);
+
+		// Adds the rotation amount based on the current orientation.
+		void OrientedRotate(const glm::vec3 amount);
+
+		glm::quat value;
+		glm::vec3 rotation;
+	};
+}
