--- src/transform.cpp
@@ -1,5 +1,0 @@
-#include "transform.hpp"
-
-namespace tec {
-
-}
